take PS=-1 into account in CTPSimulation and set prescale mode on data to RANDOM by default (TrigT1CTP-03-00-81)
	* take PS=-1 into account in CTPSimulation and set prescale mode on data to RANDOM by default
	* TrigT1CTP-03-00-81

2016-04-13  Thorsten Wengler  <thorsten.wengler@cern.ch>

	* Fix compiler warning
	* TrigT1CTP-03-00-80

+++ Trigger/TrigT1/TrigT1CTP/TrigT1CTP/CTPSimulation.h
@@ -196,7 +196,7 @@ namespace LVL1CTP {
       ResultBuilder* m_resultBuilder;                         //!< handle to result builder (for easy data access)
 
       // additional monitoring variables
-      std::vector<int> m_prescales;                           //!< prescale set (for monitoring)
+      std::vector<float> m_prescales;                         //!< prescale set (for monitoring)
 
       std::vector<int> m_countsBP;                            //!< Counts for summary BP
       std::vector<int> m_countsAP;                            //!< Counts for summary AP
